+++ docs/source/release_notes.rst
@@ -6,6 +6,7 @@ Development - |version|
 .. *Release Date: MMM. DD, YYYY*
 * Add an example script where reward is based on the probability of successfully observing targets covered 
   by clouds in the `Cloud Environment with Re-imaging <examples/cloud_environment_with_reimaging.ipynb>`_ example.
+* Add a conjunction checking dynamics model in :class:`~bsk_rl.dynamics.ConjunctionDynModel`.
 
 Version 1.1.0
 -------------